function to userID
